requirements.txt
gradio
transformers>=4.45.0
peft>=0.7.0
accelerate
bitsandbytes
torch
Pillow
huggingface_hub
